requirements.txt
Faker
Flask==3.0.2
google-ai-generativelanguage==0.6.10
google-api-core==2.17.0
google-api-python-client==2.122.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-generativeai==0.8.3
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-interceptor==0.15.3
grpcio==1.67.0
grpcio-status==1.62.3
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==1.5.3
pandas-gbq==0.23.1
psycopg2-binary==2.9.9
psycopg2-pool==1.1
SQLAlchemy==1.4.50
